src/main/java/leetcode_categories/binary_tree/L94BinaryTreeInorderTraversal.java
package leetcode_categories.binary_tree;

import utils.TreeNode;

import java.util.*;

public class L94BinaryTreeInorderTraversal {
    public List<Integer> inorderTraversal(TreeNode root) {
        Stack<TreeNode> stack = new Stack<>();
        List<Integer> result = new ArrayList<>();
        while (root != null || !stack.isEmpty()) {
            if(root != null) {
                stack.add(root);
                root = root.left;
            }
            else {
                root = stack.pop();
                result.add(root.val);
                root = root.right;
            }
        }
        return result;
    }
}
